requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
djangorestframework==3.15.2
idna==3.7
markdownify==0.11.6
numpy==1.26.3
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-jobspy==1.1.60
pytz==2024.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
tls-client==1.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
